[level 1] Title: 실패율, Time: 9.79 ms, Memory: 100 MB -BaekjoonHub

+++ 프로그래머스/1/42889. 실패율/실패율.java
@@ -15,7 +15,7 @@ class Solution {
             if(challenger[i] == 0) {
                 fails.put(i, 0.);
             } else {
-                fails.put(i, challenger[i] / total);
+                fails.put(i, (double) challenger[i] / total);
                 total -= challenger[i];
             }
         }
